+++ tests/e2e/test_toplevel.py
@@ -5,6 +5,7 @@ from uuid import uuid4
 
 import pandas as pd
 import polars as pl
+import pytest
 
 import nominal as nm
 from nominal import _utils
@@ -180,6 +181,34 @@ def test_search_runs():
     assert run2.end == run.end == nm.ts._SecondsNanos.from_datetime(end).to_nanoseconds()
     assert run2.properties == run.properties == {}
     assert run2.labels == run.labels == ()
+
+
+def test_search_runs_substring():
+    name = f"run-{uuid4()}"
+    desc = f"top-level test to search for a run by name {uuid4()}"
+    start, end = _create_random_start_end()
+    run = nm.create_run(name, start, end, desc)
+    runs = nm.search_runs(name_substring=name[4:])
+    assert len(runs) == 1
+    run2 = runs[0]
+
+    assert run2.rid == run.rid != ""
+    assert run2.name == run.name == name
+
+
+def test_search_runs_substring_deprecated():
+    name = f"run-{uuid4()}"
+    desc = f"top-level test to search for a run by name {uuid4()}"
+    start, end = _create_random_start_end()
+    run = nm.create_run(name, start, end, desc)
+
+    with pytest.warns(UserWarning):
+        runs = nm.search_runs(exact_name=name[4:])
+    assert len(runs) == 1
+    run2 = runs[0]
+
+    assert run2.rid == run.rid != ""
+    assert run2.name == run.name == name
 
 
 def test_upload_attachment(csv_data):